pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maestro-tempo"
version = "0.1.0"
description = "Zero-gas agent orchestrator for Tempo. Scoped session keys, fee-sponsored parallel execution, atomic batch settlement."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pympp[tempo]>=0.5",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
maestro = "maestro.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["maestro"]
